requirements-dev.txt
pytest==8.3.5
pyyaml>=6.0
